Refactor table component 🧰
